requirements.txt
# Discord
discord.py>=2.3.0

# Database
SQLAlchemy>=2.0.0
alembic>=1.12.0

# Environment variables
python-dotenv>=1.0.0

# PostgreSQL support (optional)
psycopg2-binary>=2.9.0

# Utilities
python-dateutil>=2.8.0